Move WebSurfer Autogen code out of the studio module
