+++ docs/docs/en/api/fastagency/runtime/autogen/base/WebSurferChat.md
@@ -1,0 +1,11 @@
+---
+# 0.5 - API
+# 2 - Release
+# 3 - Contributing
+# 5 - Template Page
+# 10 - Default
+search:
+  boost: 0.5
+---
+
+::: fastagency.runtime.autogen.base.WebSurferChat
